Update client.py to add scheduled_time_at_destination parameter
Added a new "scheduled_time_at_destination" parameter, that can be used to show what time the train was due to arrive at the destination as per the timetable. This allows you to then use this parameter in a card to show scheduled time of arrival vs estimated time of arrival.